added movies list to ui

+++ app/src/main/java/com/interview/promova_app/data/local/MovieEntity.kt
@@ -10,5 +10,5 @@ data class MovieEntity(
     val poster_path: String,
     val release_date: String,
     val title: String,
-    val vote_average: Float,
+    val vote_average: Double,
 )
